Split off mutable builder from CredentialLoader.
This is an attempt to fix a flaky test. I still do not have a good
explanation for the flake other than optimizer running amok, but
this code restructuring will make it easier to reason about the code,
as mutable and immutable objects are now split apart. It may also
make it easier for the optimizer to reason about the code making
the flake go away.
